dbsql-with-aad-token/requirements.txt
azure-identity
databricks-sql-connector
